+++ CHANGELOG.md
@@ -19,6 +19,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Support generating accessors for generic types ([#10](https://github.com/nesk/akkurate/issues/10))
 - A KDoc is provided for each validatable accessor ([#6](https://github.com/nesk/akkurate/issues/6))
 
+### Fixed
+
+- Validatable accessors are no longer generated for extension properties 
+
 ### Changed
 
 - Propagate the `@Validate` annotation to the nested classes ([#11](https://github.com/nesk/akkurate/issues/11))